requirements.txt
PyGithub>=2.1.0
anthropic>=0.18.0
python-docx>=1.1.0
requests>=2.31.0
